Added more testing, and memory wrapping

+++ inc/Nes/Nes.h
@@ -17,7 +17,7 @@ public:
 
     public:
         void writeByte(uint32_t address, uint8_t data);
-        void writeWord(uint32_t address, uint8_t data);
+        void writeWord(uint32_t address, uint16_t data);
         uint8_t readByte(uint32_t address) const;
         uint16_t readWord(uint32_t address) const;
 #endif
